# Replace boundary on StringCharacterIndexPrimitiveNode