@chrisseaton 2021-06-22
`reg` is within margin of error, right?