@eregon 2021-06-25
The 2 comments above are not addressed yet.
@LillianZ do you want to address them in this PR?
They can also be addressed later, but we shouldn't forget about it (the first one is a missed optimization, the second is purely cleanup; I think @nirvdrum was looking into that one).